requirements.txt
flask
flask-sqlalchemy
flask-login
flask-wtf
pillow
flask-cors
email-validator